requirements.txt
alembic==1.5.7
aniso8601==9.0.1
attrs==20.3.0
bcrypt==3.2.0
boto3==1.18.14
botocore==1.21.14
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.1.0
flask-marshmallow==0.11.0
Flask-Migrate==2.7.0
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
glob2==0.7
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.0.0
gspread==3.7.0
httplib2==0.19.1
idna==2.10
importlib-metadata==3.7.3
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
numpy==1.19.5
oauth2-client==1.2.1
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pandas==1.1.5
protobuf==3.15.8
psycopg2-binary==2.8.6
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
sidetable==0.8.0
six==1.15.0
SQLAlchemy==1.4.2
SQLAlchemy-Utils==0.36.8
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.4.5
zipp==3.4.1
